fixes for path

+++ .github/actions/win/win_build_script.py
@@ -2,12 +2,11 @@ import os
 import shutil
 
 github_workspace = os.getenv('GITHUB_WORKSPACE')
-
 base_lib_dir = os.path.join(github_workspace, 'out/full_deploy/host')
-
 base_bin_dir = os.path.join(github_workspace, 'out/full_deploy/host/ffmpeg')
-
 target_dir = os.path.join(github_workspace, 'out/windows')
+
+os.makedirs(target_dir, exist_ok=True)    
 
 files_to_copy = [
     "ffmpeg.exe",
@@ -27,8 +26,6 @@ files_to_copy = [
     "vorbisenc.dll",
     "vorbisfile.dll"
 ]
-
-os.makedirs(target_dir, exist_ok=True)
 
 def find_binary_file(binary_name):
     for root, dirs, files in os.walk(base_bin_dir):
@@ -50,10 +47,10 @@ for filename in files_to_copy:
     else:
         source_path = find_library_file(filename) 
 
-    if os.path.isfile(source_path):
+    if source_path is not None and os.path.isfile(source_path):
         shutil.copy2(source_path, target_dir)
         print(f"Copied {source_path} to {target_dir}")
     else:
-        print(f"File not found: {source_path}")
+        print(f"File not found: {filename} (resolved path: {source_path})")
 
 print("File copying completed.")
